mingw: use uintptr_t when converting a pointer to an int
win64 uses 32 bit long, so we cannot use a long to hold a 64 bit
pointer. Thankfully, there's a [u]intptr_t type available exactly
for these uses.

+++ client/canvas.cpp
@@ -70,7 +70,7 @@ void Canvas::clear()
 
 void Canvas::begin_draw(SpiceMsgDisplayBase& base, int size, size_t min_size)
 {
-    _base = (unsigned long)&base;
+    _base = (uintptr_t)&base;
 }
 
 void Canvas::draw_fill(SpiceMsgDisplayDrawFill& fill, int size)
